# autogenerated/go-client/api/openapi.yaml
openapi: 3.0.1
info:
  contact:
    email: open-service-broker-api@googlegroups.com
    name: Open Service Broker API
    url: https://www.openservicebrokerapi.org/
  description: The Open Service Broker API defines an HTTP(S) interface between Platforms
    and Service Brokers.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Open Service Broker API
  version: master - might contain changes that are not yet released
...
servers:
- url: http://localhost:3000/
- url: https://localhost:3000/
security:
- basicAuth: []
paths:
...
  /v2/service_instances/{instance_id}:
    delete:
      operationId: serviceInstance.deprovision
      parameters:
      - description: version number of the Service Broker API that the Platform will
          use
        in: header
        name: X-Broker-API-Version
        required: true
        schema:
          type: string
      - description: identity of the user that initiated the request from the Platform
        in: header
        name: X-Broker-API-Originating-Identity
        schema:
          type: string
      - description: idenity of the request from the Platform
        in: header
        name: X-Broker-API-Request-Identity
        schema:
          type: string
      - description: instance id of instance to provision
        in: path
        name: instance_id
        required: true
        schema:
          type: string
      - description: asynchronous operations supported
        in: query
        name: accepts_incomplete
        schema:
          type: boolean
      - description: id of the service associated with the instance being deleted
        in: query
        name: service_id
        required: true
        schema:
          type: string
      - description: id of the plan associated with the instance being deleted
        in: query
        name: plan_id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
          description: OK
        "202":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsyncOperation'
          description: Accepted
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "410":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Gone
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Precondition Failed
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unprocessable Entity
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected
      summary: deprovision a service instance
      tags:
      - ServiceInstances
    get:
      operationId: serviceInstance.get
      parameters:
      - description: version number of the Service Broker API that the Platform will
          use
        in: header
        name: X-Broker-API-Version
        required: true
        schema:
          type: string
      - description: identity of the user that initiated the request from the Platform
        in: header
        name: X-Broker-API-Originating-Identity
        schema:
          type: string
      - description: idenity of the request from the Platform
        in: header
        name: X-Broker-API-Request-Identity
        schema:
          type: string
      - description: instance id of instance to provision
        in: path
        name: instance_id
        required: true
        schema:
          type: string
      - description: id of the service associated with the instance
        in: query
        name: service_id
        schema:
          type: string
      - description: id of the plan associated with the instance
        in: query
        name: plan_id
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceInstanceResource'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Not Found
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Precondition Failed
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected
      summary: gets a service instance
      tags:
      - ServiceInstances
    patch:
      operationId: serviceInstance.update
      parameters:
      - description: version number of the Service Broker API that the Platform will
          use
        in: header
        name: X-Broker-API-Version
        required: true
        schema:
          type: string
      - description: identity of the user that initiated the request from the Platform
        in: header
        name: X-Broker-API-Originating-Identity
        schema:
          type: string
      - description: idenity of the request from the Platform
        in: header
        name: X-Broker-API-Request-Identity
        schema:
          type: string
      - description: instance id of instance to provision
        in: path
        name: instance_id
        required: true
        schema:
          type: string
      - description: asynchronous operations supported
        in: query
        name: accepts_incomplete
        schema:
          type: boolean
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceInstanceUpdateRequest'
        description: parameters for the requested service instance update
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
          description: OK
        "202":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceInstanceAsyncOperation'
          description: Accepted
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Precondition Failed
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unprocessable Entity
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected
      summary: update a service instance
      tags:
      - ServiceInstances
      x-codegen-request-body-name: body
    put:
      operationId: serviceInstance.provision
      parameters:
      - description: version number of the Service Broker API that the Platform will
          use
        in: header
        name: X-Broker-API-Version
        required: true
        schema:
          type: string
      - description: identity of the user that initiated the request from the Platform
        in: header
        name: X-Broker-API-Originating-Identity
        schema:
          type: string
      - description: idenity of the request from the Platform
        in: header
        name: X-Broker-API-Request-Identity
        schema:
          type: string
      - description: instance id of instance to provision
        in: path
        name: instance_id
        required: true
        schema:
          type: string
      - description: asynchronous operations supported
        in: query
        name: accepts_incomplete
        schema:
          type: boolean
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceInstanceProvisionRequest'
        description: parameters for the requested service instance provision
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceInstanceProvisionResponse'
          description: OK
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceInstanceProvisionResponse'
          description: Created
        "202":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceInstanceAsyncOperation'
          description: Accepted
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Conflict
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Precondition Failed
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unprocessable Entity
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected
      summary: provision a service instance
      tags:
      - ServiceInstances
      x-codegen-request-body-name: body
...
  /v2/service_instances/{instance_id}/service_bindings/{binding_id}:
    delete:
      operationId: serviceBinding.unbinding
      parameters:
      - description: version number of the Service Broker API that the Platform will
          use
        in: header
        name: X-Broker-API-Version
        required: true
        schema:
          type: string
      - description: identity of the user that initiated the request from the Platform
        in: header
        name: X-Broker-API-Originating-Identity
        schema:
          type: string
      - description: idenity of the request from the Platform
        in: header
        name: X-Broker-API-Request-Identity
        schema:
          type: string
      - description: instance id of instance to provision
        in: path
        name: instance_id
        required: true
        schema:
          type: string
      - description: binding id of binding to create
        in: path
        name: binding_id
        required: true
        schema:
          type: string
      - description: id of the service associated with the instance being deleted
        in: query
        name: service_id
        required: true
        schema:
          type: string
      - description: id of the plan associated with the instance being deleted
        in: query
        name: plan_id
        required: true
        schema:
          type: string
      - description: asynchronous operations supported
        in: query
        name: accepts_incomplete
        schema:
          type: boolean
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
          description: OK
        "202":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsyncOperation'
          description: Accepted
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "410":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Gone
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Precondition Failed
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unprocessable Entity
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected
      summary: deprovision of a service binding
      tags:
      - ServiceBindings
    get:
      operationId: serviceBinding.get
      parameters:
      - description: version number of the Service Broker API that the Platform will
          use
        in: header
        name: X-Broker-API-Version
        required: true
        schema:
          type: string
      - description: identity of the user that initiated the request from the Platform
        in: header
        name: X-Broker-API-Originating-Identity
        schema:
          type: string
      - description: idenity of the request from the Platform
        in: header
        name: X-Broker-API-Request-Identity
        schema:
          type: string
      - description: instance id of instance to provision
        in: path
        name: instance_id
        required: true
        schema:
          type: string
      - description: binding id of binding to create
        in: path
        name: binding_id
        required: true
        schema:
          type: string
      - description: id of the service associated with the instance
        in: query
        name: service_id
        schema:
          type: string
      - description: id of the plan associated with the instance
        in: query
        name: plan_id
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceBindingResource'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Not Found
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Precondition Failed
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected
      summary: gets a service binding
      tags:
      - ServiceBindings
    put:
      operationId: serviceBinding.binding
      parameters:
      - description: version number of the Service Broker API that the Platform will
          use
        in: header
        name: X-Broker-API-Version
        required: true
        schema:
          type: string
      - description: identity of the user that initiated the request from the Platform
        in: header
        name: X-Broker-API-Originating-Identity
        schema:
          type: string
      - description: idenity of the request from the Platform
        in: header
        name: X-Broker-API-Request-Identity
        schema:
          type: string
      - description: instance id of instance to provision
        in: path
        name: instance_id
        required: true
        schema:
          type: string
      - description: binding id of binding to create
        in: path
        name: binding_id
        required: true
        schema:
          type: string
      - description: asynchronous operations supported
        in: query
        name: accepts_incomplete
        schema:
          type: boolean
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceBindingRequest'
        description: parameters for the requested service binding
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceBindingResponse'
          description: OK
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceBindingResponse'
          description: Created
        "202":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsyncOperation'
          description: Accepted
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Conflict
        "410":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Gone
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Precondition Failed
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unprocessable Entity
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected
      summary: generation of a service binding
      tags:
      - ServiceBindings
      x-codegen-request-body-name: body
components:
...
  schemas:
...
    ServiceInstanceResource:
      example:
        metadata:
          attributes: '{}'
          labels: '{}'
        maintenance_info:
          description: description
          version: version
        service_id: service_id
        dashboard_url: dashboard_url
        parameters: '{}'
        plan_id: plan_id
      properties:
        service_id:
          type: string
        plan_id:
          type: string
        dashboard_url:
          type: string
        parameters:
          properties: {}
          type: object
        maintenance_info:
          $ref: '#/components/schemas/MaintenanceInfo'
        metadata:
          $ref: '#/components/schemas/ServiceInstanceMetadata'
      type: object
    ServiceInstanceProvisionRequest:
      example:
        maintenance_info:
          description: description
          version: version
        service_id: service_id
        context: '{}'
        organization_guid: organization_guid
        space_guid: space_guid
        parameters: '{}'
        plan_id: plan_id
      properties:
        service_id:
          type: string
        plan_id:
          type: string
        context:
          description: See [Context Conventions](https://github.com/openservicebrokerapi/servicebroker/blob/master/profile.md#context-object)
            for more details.
          type: object
        organization_guid:
          type: string
          x-deprecated: true
        space_guid:
          type: string
          x-deprecated: true
        parameters:
          properties: {}
          type: object
        maintenance_info:
          $ref: '#/components/schemas/MaintenanceInfo'
      required:
      - organization_guid
      - plan_id
      - service_id
      - space_guid
      type: object
    ServiceInstanceProvisionResponse:
      example:
        metadata:
          attributes: '{}'
          labels: '{}'
        dashboard_url: dashboard_url
      properties:
        dashboard_url:
          type: string
        metadata:
          $ref: '#/components/schemas/ServiceInstanceMetadata'
      type: object
    ServiceInstanceAsyncOperation:
      properties:
        dashboard_url:
          type: string
        operation:
          type: string
        metadata:
          $ref: '#/components/schemas/ServiceInstanceMetadata'
      type: object
    ServiceInstanceUpdateRequest:
      example:
        maintenance_info:
          description: description
          version: version
        previous_values:
          maintenance_info:
            description: description
            version: version
          service_id: service_id
          organization_id: organization_id
          space_id: space_id
          plan_id: plan_id
        service_id: service_id
        context: '{}'
        parameters: '{}'
        plan_id: plan_id
      properties:
        context:
          description: See [Context Conventions](https://github.com/openservicebrokerapi/servicebroker/blob/master/profile.md#context-object)
            for more details.
          type: object
        service_id:
          type: string
        plan_id:
          type: string
        parameters:
          properties: {}
          type: object
        previous_values:
          $ref: '#/components/schemas/ServiceInstancePreviousValues'
        maintenance_info:
          $ref: '#/components/schemas/MaintenanceInfo'
      required:
      - service_id
      type: object
    ServiceInstancePreviousValues:
      example:
        maintenance_info:
          description: description
          version: version
        service_id: service_id
        organization_id: organization_id
        space_id: space_id
        plan_id: plan_id
      properties:
        service_id:
          type: string
          x-deprecated: true
        plan_id:
          type: string
        organization_id:
          type: string
          x-deprecated: true
        space_id:
          type: string
          x-deprecated: true
        maintenance_info:
          $ref: '#/components/schemas/MaintenanceInfo'
      type: object
    AsyncOperation:
      properties:
        operation:
          type: string
      type: object
...
    ServiceBindingResource:
      example:
        metadata:
          expires_at: expires_at
          renew_before: renew_before
        route_service_url: route_service_url
        endpoints:
        - protocol: tcp
          host: host
          ports:
          - ports
          - ports
        - protocol: tcp
          host: host
          ports:
          - ports
          - ports
        credentials: '{}'
        syslog_drain_url: syslog_drain_url
        volume_mounts:
        - mode: r
          driver: driver
          device_type: shared
          device:
            volume_id: volume_id
            mount_config: '{}'
          container_dir: container_dir
        - mode: r
          driver: driver
          device_type: shared
          device:
            volume_id: volume_id
            mount_config: '{}'
          container_dir: container_dir
        parameters: '{}'
      properties:
        metadata:
          $ref: '#/components/schemas/ServiceBindingMetadata'
        credentials:
          properties: {}
          type: object
        syslog_drain_url:
          type: string
        route_service_url:
          type: string
        volume_mounts:
          items:
            $ref: '#/components/schemas/ServiceBindingVolumeMount'
          type: array
        endpoints:
          items:
            $ref: '#/components/schemas/ServiceBindingEndpoint'
          type: array
        parameters:
          properties: {}
          type: object
      type: object
    ServiceBindingRequest:
      example:
        app_guid: app_guid
        predecessor_binding_id: predecessor_binding_id
        service_id: service_id
        context: '{}'
        bind_resource:
          app_guid: app_guid
          route: route
        parameters: '{}'
        plan_id: plan_id
      properties:
        context:
          description: See [Context Conventions](https://github.com/openservicebrokerapi/servicebroker/blob/master/profile.md#context-object)
            for more details.
          type: object
        service_id:
          type: string
        plan_id:
          type: string
        app_guid:
          type: string
          x-deprecated: true
        bind_resource:
          $ref: '#/components/schemas/ServiceBindingResourceObject'
        parameters:
          properties: {}
          type: object
        predecessor_binding_id:
          type: string
      required:
      - plan_id
      - service_id
      type: object
...
    ServiceBindingResponse:
      example:
        metadata:
          expires_at: expires_at
          renew_before: renew_before
        route_service_url: route_service_url
        endpoints:
        - protocol: tcp
          host: host
          ports:
          - ports
          - ports
        - protocol: tcp
          host: host
          ports:
          - ports
          - ports
        credentials: '{}'
        syslog_drain_url: syslog_drain_url
        volume_mounts:
        - mode: r
          driver: driver
          device_type: shared
          device:
            volume_id: volume_id
            mount_config: '{}'
          container_dir: container_dir
        - mode: r
          driver: driver
          device_type: shared
          device:
            volume_id: volume_id
            mount_config: '{}'
          container_dir: container_dir
      properties:
        metadata:
          $ref: '#/components/schemas/ServiceBindingMetadata'
        credentials:
          properties: {}
          type: object
        syslog_drain_url:
          type: string
        route_service_url:
          type: string
        volume_mounts:
          items:
            $ref: '#/components/schemas/ServiceBindingVolumeMount'
          type: array
        endpoints:
          items:
            $ref: '#/components/schemas/ServiceBindingEndpoint'
          type: array
      type: object
    ServiceBindingMetadata:
      example:
        expires_at: expires_at
        renew_before: renew_before
      properties:
        expires_at:
          type: string
        renew_before:
          type: string
      type: object
...
    ServiceBindingVolumeMount:
      example:
        mode: r
        driver: driver
        device_type: shared
        device:
          volume_id: volume_id
          mount_config: '{}'
        container_dir: container_dir
      properties:
        driver:
          type: string
        container_dir:
          type: string
        mode:
          enum:
          - r
          - rw
          type: string
        device_type:
          enum:
          - shared
          type: string
        device:
          $ref: '#/components/schemas/ServiceBindingVolumeMountDevice'
      required:
      - container_dir
      - device
      - device_type
      - driver
      - mode
      type: object
...
    ServiceInstanceMetadata:
      example:
        attributes: '{}'
        labels: '{}'
      properties:
        labels:
          properties: {}
          type: object
        attributes:
          properties: {}
          type: object
      type: object
    MaintenanceInfo:
      example:
        description: description
        version: version
      properties:
        version:
          type: string
        description:
          type: string
      type: object
...
    Error:
      properties:
        error:
          type: string
        description:
          type: string
        instance_usable:
          type: boolean
        update_repeatable:
          type: boolean
      type: object
  securitySchemes:
    basicAuth:
      scheme: basic
      type: http
...